% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscale_spec.R
\name{gtscale_apply}
\alias{gtscale_apply}
\title{Apply only the color component of a \code{gtscales} spec}
\usage{
gtscale_apply(data, spec)
}
\arguments{
\item{data}{A \code{gt_tbl} created by \code{\link[gt:gt]{gt::gt()}}.}

\item{spec}{A \code{gtscale_spec}.}
}
\value{
A modified \link[gt:gt]{gt::gt} table.
}
\description{
Apply only the color component of a \code{gtscales} spec
}
