% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscale_spec.R
\name{gtscale_apply_legend}
\alias{gtscale_apply_legend}
\title{Apply a \code{gtscales} spec and attach its legend}
\usage{
gtscale_apply_legend(data, spec)
}
\arguments{
\item{data}{A \code{gt_tbl} created by \code{\link[gt:gt]{gt::gt()}}.}

\item{spec}{A \code{gtscale_spec}.}
}
\value{
A modified \link[gt:gt]{gt::gt} table.
}
\description{
Apply a \code{gtscales} spec and attach its legend
}
\examples{
library(gt)

spec <- gtscale_spec_continuous(
  num,
  palette = c('#A0442C', 'white', '#0063B1'),
  title = 'Value'
)

exibble |>
  gt() |>
  gtscale_apply_legend(spec)
}
