% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_numeric.R
\name{gtscale_color_continuous}
\alias{gtscale_color_continuous}
\title{Add a scale for a numeric column}
\usage{
gtscale_color_continuous(data, fn)
}
\arguments{
\item{data}{A \code{gt} table with class \code{gt_tbl}}

\item{fn}{a function from \code{scales}}
}
\value{
A \link[gt:gt]{gt::gt}
}
\description{
Add a scale for a numeric column
}
\examples{
party_purple <- scales::col_numeric(
  domain = c(0, 1),
  palette = c('#A0442C', '#0063B1')
)
}
