% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_numeric.R
\name{gtscale_color_numeric}
\alias{gtscale_color_numeric}
\title{Add a scale for a numeric column}
\usage{
gtscale_color_numeric(data, fn)
}
\arguments{
\item{data}{A \code{gt} table with class \code{gt_tbl}}

\item{fn}{a function from \code{scales}}
}
\value{
A \link[gt:gt]{gt::gt}
}
\description{
Add a scale for a numeric column
}
\examples{
#TODO
}
