% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscale_color_quantiles.R
\name{gtscale_color_quantiles}
\alias{gtscale_color_quantiles}
\title{Add only a quantile color legend to a \code{gt} table}
\usage{
gtscale_color_quantiles(
  data,
  column,
  palette,
  quantiles = 4,
  labels = NULL,
  title = NULL,
  width = "180px",
  height = "14px"
)
}
\arguments{
\item{data}{A \code{gt_tbl} created by \code{\link[gt:gt]{gt::gt()}}.}

\item{column}{A numeric column in the underlying table used to infer quantile
boundaries.}

\item{palette}{A vector of colors or palette endpoints used to color the
quantile bins.}

\item{quantiles}{The number of quantile groups.}

\item{labels}{A labeling function or a character vector for the quantile
ranges. When a function is supplied, it is applied to the quantile
boundaries before interval labels are constructed.}

\item{title}{Optional legend title.}

\item{width}{Width of the legend.}

\item{height}{Height of the swatches.}
}
\value{
A modified \link[gt:gt]{gt::gt} table.
}
\description{
This is a lower-level helper for cases where table coloring is already
handled elsewhere. For the usual "color and legendize" workflow, prefer
\code{\link[=gtscale_data_color_quantiles]{gtscale_data_color_quantiles()}}.
}
\examples{
library(gt)

exibble |>
  gt() |>
  data_color(
    columns = num,
    method = 'quantile',
    palette = c('#fdd49e', '#fdbb84', '#ef6548', '#990000'),
    quantiles = 4
  ) |>
  gtscale_color_quantiles(
    column = num,
    palette = c('#fdd49e', '#fdbb84', '#ef6548', '#990000'),
    quantiles = 4,
    title = 'Quantile bins'
  )
}
