% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscale_spec.R
\name{gtscale_spec_bins}
\alias{gtscale_spec_bins}
\title{Create a binned \code{gtscales} spec}
\usage{
gtscale_spec_bins(
  column,
  palette,
  bins = NULL,
  domain = NULL,
  transform = NULL,
  oob = NULL,
  right = FALSE,
  labels = NULL,
  title = NULL,
  width = "180px",
  height = "14px"
)
}
\arguments{
\item{column}{A column to target.}

\item{palette}{A vector of colors or palette endpoints used for the bins.
A single named palette or palette function can also be supplied.}

\item{bins}{Optional bin boundaries or a break function. When omitted,
breaks are generated from the \code{domain}, \code{column}, and \code{transform}.}

\item{domain}{Optional limits. If omitted, these can be inferred when the
spec is applied to a \code{gt} table.}

\item{transform}{A transformation specification understood by
\code{\link[scales:new_transform]{scales::as.transform()}}. This is used when generating default bins or
when interpreting break functions.}

\item{oob}{Out-of-bounds handling function or shortcut. Use a function like
\code{\link[scales:oob]{scales::oob_squish()}} or a shortcut such as \code{"censor"} or \code{"squish"}.}

\item{right}{Whether intervals should be closed on the right. The default of
\code{FALSE} yields intervals like \verb{[a, b)}.}

\item{labels}{An optional labeling function or character vector for the
legend. When omitted, labels are inferred from the bin values.}

\item{title}{Optional legend title.}

\item{width}{Width of the legend.}

\item{height}{Height of the swatches.}
}
\value{
A \code{gtscale_spec}.
}
\description{
Create a binned \code{gtscales} spec
}
