% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscale_spec.R
\name{gtscale_spec_discrete}
\alias{gtscale_spec_discrete}
\title{Create a discrete \code{gtscales} spec}
\usage{
gtscale_spec_discrete(
  column,
  values,
  labels = values,
  title = NULL,
  swatch_size = "12px",
  levels = NULL,
  ordered = FALSE
)
}
\arguments{
\item{column}{A column to target.}

\item{values}{A vector of color values, a single named discrete palette, or a
discrete palette function.}

\item{labels}{Labels for each legend swatch. Defaults to \code{values}.}

\item{title}{Optional legend title.}

\item{swatch_size}{Size of each discrete color swatch.}

\item{levels}{Optional factor levels.}

\item{ordered}{Whether the scale should be treated as ordered.}
}
\value{
A \code{gtscale_spec}.
}
\description{
Create a discrete \code{gtscales} spec
}
