% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscale_spec.R
\name{gtscale_spec_set_application}
\alias{gtscale_spec_set_application}
\title{Set how a \code{gtscales} spec is applied}
\usage{
gtscale_spec_set_application(
  spec,
  apply_to = c("fill", "text"),
  na_color = NULL,
  alpha = NULL,
  reverse = FALSE,
  accessibility = c("none", "warn"),
  autocolor_text = TRUE,
  contrast_algo = c("apca", "wcag"),
  autocolor_light = "#FFFFFF",
  autocolor_dark = "#000000"
)
}
\arguments{
\item{spec}{A \code{gtscale_spec}.}

\item{apply_to}{Whether colors should be applied to cell fill or text.}

\item{na_color}{Color used for missing values.}

\item{alpha}{Alpha applied by \code{\link[gt:data_color]{gt::data_color()}}.}

\item{reverse}{Whether to reverse the color mapping.}

\item{accessibility}{Whether to warn about low-contrast adjacent legend
colors.}

\item{autocolor_text}{Whether to automatically adjust text color.}

\item{contrast_algo}{Contrast algorithm passed to \code{\link[gt:data_color]{gt::data_color()}}.}

\item{autocolor_light}{Light text color used by \code{\link[gt:data_color]{gt::data_color()}}.}

\item{autocolor_dark}{Dark text color used by \code{\link[gt:data_color]{gt::data_color()}}.}
}
\value{
A modified \code{gtscale_spec}.
}
\description{
Set how a \code{gtscales} spec is applied
}
