% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscale_spec.R
\name{gtscale_spec_set_legend}
\alias{gtscale_spec_set_legend}
\title{Set how a \code{gtscales} legend should be rendered}
\usage{
gtscale_spec_set_legend(spec, output = "html", placement = "source_note")
}
\arguments{
\item{spec}{A \code{gtscale_spec}.}

\item{output}{Output target for the legend. Currently only \code{"html"} is
implemented.}

\item{placement}{Legend placement target. Currently only \code{"source_note"} is
implemented.}
}
\value{
A modified \code{gtscale_spec}.
}
\description{
Set how a \code{gtscales} legend should be rendered
}
