% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscale_spec.R
\name{gtscale_spec_continuous}
\alias{gtscale_spec_continuous}
\title{Create a continuous \code{gtscales} spec}
\usage{
gtscale_spec_continuous(
  column,
  palette = NULL,
  domain = NULL,
  breaks = NULL,
  labels = scales::label_comma(),
  title = NULL,
  direction = "to right",
  width = "160px",
  height = "14px",
  fn = NULL
)
}
\arguments{
\item{column}{A column to target.}

\item{palette}{A vector of colors used in the scale.}

\item{domain}{Optional numeric limits. If omitted, these can be inferred when
the spec is applied to a \code{gt} table.}

\item{breaks}{Optional numeric break values for the legend.}

\item{labels}{A labeling function or character vector for the legend.}

\item{title}{Optional legend title.}

\item{direction}{CSS gradient direction. Defaults to \code{"to right"}.}

\item{width}{Width of the legend bar.}

\item{height}{Height of the legend bar.}

\item{fn}{Optional \code{scales} function for numeric coloring.}
}
\value{
A \code{gtscale_spec}.
}
\description{
Create a continuous \code{gtscales} spec
}
