% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtscales-package.R
\docType{package}
\name{gtscales-package}
\alias{gtscales}
\alias{gtscales-package}
\title{gtscales: Color Scale Extensions for 'gt'}
\description{
Adds legend helpers for color-encoded 'gt' tables. The primary interface combines 'gt::data_color()' with matched legends for continuous, binned, quantile, and discrete scales, while lower-level helpers support legend-only and reusable scale specification workflows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://christophertkenny.com/gtscales/}
}

}
\author{
\strong{Maintainer}: Christopher T. Kenny \email{ctkenny@proton.me} (\href{https://orcid.org/0000-0002-9386-6860}{ORCID})

}
\keyword{internal}
